euro_lottery/backend/requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
Django==4.2.9
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework_simplejwt==5.3.0
idna==3.6
pillow==10.2.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-dotenv==1.0.0
requests==2.31.0
sqlparse==0.4.4
stripe==7.13.0
typing_extensions==4.9.0
urllib3==2.1.0
gunicorn==21.2.0
whitenoise==6.6.0
celery==5.3.6
redis==5.0.1
django-celery-beat==2.5.0
drf-yasg==1.21.7
pytest==7.4.3
pytest-django==4.5.2
coverage==7.3.2
black==23.3.0
flake8==6.0.0
isort==5.12.0
python-json-logger==2.0.7
